+++ src/routes/ProgramsScreen/ProgramsScreen.tsx
@@ -9,20 +9,83 @@ import Category from 'types/Category'
 import { programsScreenLabels } from 'labels'
 
 import Params from './Params'
-import ClipPrograms from './ClipPrograms'
+import AudioPrograms from './AudioPrograms'
 import VideoPrograms from './VideoPrograms'
 import { programsScreen, programTypeNav, typeItem, active } from './ProgramsScreen.scss'
 
-type ProgramType = 'clip' | 'gallery' | 'video'
+type ProgramType = 'audio' | 'video'
 
-const CLIP: ProgramType = 'clip'
-// const GALLERY: ProgramType = 'gallery'
+const AUDIO: ProgramType = 'audio'
 const VIDEO: ProgramType = 'video'
 
 const PROGRAM_ZONES: Category[] = [
   {
-    id: 0,
-    name: programsScreenLabels.all,
+    id: 1927,
+    name: '每日视频新闻',
+  },
+  {
+    id: 1737,
+    name: '学英语',
+  },
+  {
+    id: 2542,
+    name: 'VOA卫视完整版',
+  },
+  {
+    id: 4982,
+    name: '美国观察',
+  },
+  {
+    id: 1904,
+    name: '时事大家谈',
+  },
+  {
+    id: 1905,
+    name: '焦点对话',
+  },
+  {
+    id: 5327,
+    name: '周末专辑',
+  },
+  {
+    id: 1906,
+    name: '海峡论谈',
+  },
+  {
+    id: 1903,
+    name: 'VOA连线',
+  },
+  {
+    id: 3690,
+    name: '媒体观察',
+  },
+  {
+    id: 2422,
+    name: '时事看台',
+  },
+  {
+    id: 5257,
+    name: '专家视点',
+  },
+  {
+    id: 1907,
+    name: '美国万花筒',
+  },
+  {
+    id: 2421,
+    name: '走进美国',
+  },
+  {
+    id: 5256,
+    name: '鹰与盾：美国军事透视',
+  },
+  {
+    id: 4495,
+    name: '您的孩子在美国',
+  },
+  {
+    id: 5295,
+    name: '对话名校招生官',
   },
 ]
 
@@ -50,8 +113,8 @@ class ProgramsScreen extends React.Component<Props> {
   renderPrograms () {
     const { history, match } = this.props
     const { type = VIDEO } = match.params
-    if (type === CLIP) {
-      return <ClipPrograms history={history} />
+    if (type === AUDIO) {
+      return <AudioPrograms history={history} match={match} />
     } else if (type === VIDEO) {
       return <VideoPrograms history={history} match={match} />
     } else {
@@ -67,11 +130,9 @@ class ProgramsScreen extends React.Component<Props> {
         <div className={type === VIDEO ? `${typeItem} ${active}` : typeItem} onClick={() => this.setProgramType(VIDEO)}>
           {programsScreenLabels.videos}
         </div>
-        {
-        // <div className={type === CLIP ? `${typeItem} ${active}` : typeItem} onClick={() => this.setProgramType(CLIP)}>
-        //   {programsScreenLabels.audio}
-        // </div>
-        }
+        <div className={type === AUDIO ? `${typeItem} ${active}` : typeItem} onClick={() => this.setProgramType(AUDIO)}>
+          {programsScreenLabels.audio}
+        </div>
       </div>
     )
   }
@@ -108,6 +169,7 @@ class ProgramsScreen extends React.Component<Props> {
         <ErrorBoundary>
           {this.renderPrograms()}
         </ErrorBoundary>
+        {this.renderProgramTypes()}
       </div>
     )
   }